// frontend/src/api/gen/types/SidequestDto.ts
export type SidequestDto = {
  /**
   * @type string
   */
  description: string;
  /**
   * @type string, uuid
   */
  event_id: string;
  /**
   * @type string, uuid
   */
  id: string;
  /**
   * @type boolean
   */
  is_higher_result_better: boolean;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  slug: string;
};